Columnar Encryption
Col. trans. encrypt function is constructed, even with a few comments.
This has some native and necessary padding, but it's all done internally
and shouldn't mess with the additional padding we want.

I think I figured out a bubble sort (or perhaps a reverse bubble sort)
all by myself!

The decryption may be annoying on this one.